video clip implementation

+++ audio_processing/recorder.py
@@ -33,29 +33,30 @@ class WaveRecorder:
 
 class VidRecorder:
 
-    def __init__(self, filename, tempvidfilename, tempaudfilename):
-        self.dat_filename = filename + '.dat'
+    def __init__(self, filename, tempvidfilename, tempaudfilename,is_save):
+        #self.dat_filename = filename + '.dat'
         self.vid_filename = filename + '.mp4'
-        self.temp_vid_filename =  tempvidfilename + '.mp4'
-        self.temp_aud_filename =  tempaudfilename + '.wav'
+        # self.temp_vid_filename =  tempvidfilename + '.mp4'
+        # self.temp_aud_filename =  tempaudfilename + '.wav'
         self.closed = False
+        self.save_video = is_save
 
     def write(self, data):
         try:
-            with open(self.dat_filename, "ab") as f:
+            with open(self.vid_filename, "ab") as f:
                 f.write(data)
         except Exception as e:
             print('Unable to write data to file: {0}'.format(e))
 
-    def write_temp_mp4(self,data):
-        try:
-            if os.path.isfile(self.temp_vid_filename):
-                os.remove(self.temp_vid_filename)
+    # def write_temp_mp4(self,data):
+    #     try:
+    #         if os.path.isfile(self.temp_vid_filename):
+    #             os.remove(self.temp_vid_filename)
 
-            with open(self.temp_vid_filename, "wb") as out_file:
-                out_file.write(data)
-        except Exception as e:
-            print('Unable to write temp video data to file: {0}'.format(e))        
+    #         with open(self.temp_vid_filename, "wb") as out_file:
+    #             out_file.write(data)
+    #     except Exception as e:
+    #         print('Unable to write temp video data to file: {0}'.format(e))        
 
     def read_temp_wav(self):
         try:
@@ -76,14 +77,14 @@ class VidRecorder:
                     os.remove(self.temp_aud_filename)
 
                 #save dat binary file as mp4    
-                with open(self.dat_filename, "rb") as f:
-                    all_data = f.read()
+                # with open(self.dat_filename, "rb") as f:
+                #     all_data = f.read()
                 
-                if os.path.isfile(self.vid_filename):
+                if not self.save_video and  os.path.isfile(self.vid_filename):
                     os.remove(self.vid_filename)
 
-                with open(self.vid_filename, "wb") as out_file:
-                    out_file.write(all_data)  
+                # with open(self.vid_filename, "wb") as out_file:
+                #     out_file.write(all_data)  
             except Exception as e:
               print('Unable to write video data to file: {0}'.format(e))         
             os.remove(self.dat_filename)
